requirements.txt
backports.zoneinfo;python_version<"3.9"
asgiref==3.6.0
beautifulsoup4==4.12.2
Django==4.2
django-bootstrap5==23.1
django-ckeditor==6.5.1
django-cors-headers==4.1.0
django-environ==0.10.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
Pillow==9.5.0
# psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
drf-spectacular==0.26.4
django-environ


